requirement.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
chromadb
tavily-python
langgraph
pypdf
python-dotenv
streamlit